Create Home.js
 import the necessary dependencies (React, useEffect, useDispatch, useSelector) from react and redux. The Home component retrieves the book data from the Redux store using the useSelector hook. It also dispatches the fetchBooks action to fetch the books on component mount using the useEffect hook.

If the books are still loading, it renders a loading message. If there's an error, it displays an error message. Otherwise, it renders the book catalog with the help of the BookItem component. Each book is rendered using the map function and passed as a prop to the BookItem component.