Create index.js
Imported the combineReducers function from the Redux library. We also import the authReducer and reviewReducer that we previously created.

The combineReducers function is used to combine multiple reducers into a single reducer. In this case, we combine the authReducer and reviewReducer into the rootReducer.

Finally, we export the rootReducer so that it can be used in the Redux store configuration.